Start moving common functions to a module
