+++ frameworks/kafka/tests/test_zookeeper_auth.py
@@ -1,3 +1,6 @@
+"""
+This module tests the interaction of Kafka with Zookeeper with authentication enabled
+"""
 import logging
 import uuid
 import pytest
@@ -13,6 +16,7 @@ import sdk_repository
 import sdk_tasks
 import sdk_utils
 
+from tests import auth
 from tests import config
 
 log = logging.getLogger(__name__)
@@ -229,86 +233,5 @@ def kafka_client(kerberos, kafka_server):
 @sdk_utils.dcos_ee_only
 @pytest.mark.sanity
 def test_client_can_read_and_write(kafka_client):
-    wait_for_brokers(kafka_client["id"], kafka_client["brokers"])
-    send_and_receive_message(kafka_client["id"])
-
-
-@pytest.mark.dcos_min_version('1.10')
-@sdk_utils.dcos_ee_only
-@pytest.mark.sanity
-def test_client_can_read_and_write_from_kerberized_zookeeper(kafka_client,
-                                                             kafka_server,
-                                                             zookeeper_server):
-
-    # Uninstall the KAFKA client
-    client_config = sdk_marathon.get_config(kafka_client["id"])
-    sdk_marathon.destroy_app(client_config)
-
-    # Uninstall the kafka service
-    sdk_install.uninstall(config.PACKAGE_NAME, config.SERVICE_NAME)
-
-    # Get the zookeeper DNS values
-    zookeeper_dns = sdk_cmd.svc_cli(zookeeper_server["package_name"],
-                                    zookeeper_server["service"]["name"],
-                                    "endpoint clientport", json=True)["dns"]
-
-    zookeeper_options = kafka_server.copy()
-    additional_options = {
-        "service": {
-            "security": {
-                "kerberos": {
-                    "enabled_for_zookeeper": True
-                }
-            }
-        },
-        "kafka": {
-            "kafka_zookeeper_uri": ",".join(zookeeper_dns)
-        }
-    }
-    zookeeper_options.update(additional_options)
-
-    # Reinstall Kafka with zookeeper
-    sdk_install.install(
-        config.PACKAGE_NAME,
-        config.SERVICE_NAME,
-        config.DEFAULT_BROKER_COUNT,
-        additional_options=zookeeper_options,
-        timeout_seconds=30 * 60)
-
-    # Reinstall the client
-    brokers = sdk_cmd.svc_cli(config.PACKAGE_NAME,
-                              config.SERVICE_NAME,
-                              "endpoint broker", json=True)["dns"]
-    client_config["env"]["KAFKA_BROKER_LIST"] = ",".join(brokers)
-    sdk_marathon.install_app(client_config)
-
-    wait_for_brokers(kafka_client["id"], brokers)
-    send_and_receive_message(kafka_client["id"])
-
-
-def wait_for_brokers(client: str, brokers: list):
-    log.info("Running bootstrap to wait for DNS resolution")
-    bootstrap_cmd = ['/opt/bootstrap',
-                     '-resolve-hosts', ','.join(brokers), '-verbose']
-    bootstrap_output = sdk_tasks.task_exec(client, ' '.join(bootstrap_cmd))
-    log.info(bootstrap_output)
-
-
-def send_and_receive_message(client: str):
-    message = uuid.uuid4()
-    producer_cmd = ['/tmp/kafkaconfig/start.sh', 'producer', str(message)]
-
-    for i in range(2):
-        log.info("Running(%s) %s", i, producer_cmd)
-        producer_output = sdk_tasks.task_exec(client, ' '.join(producer_cmd))
-        log.info("Producer output(%s): %s", i, producer_output)
-
-    assert "Sent message: '{message}'".format(message=str(
-        message)) in ' '.join(str(p) for p in producer_output)
-
-    consumer_cmd = ['/tmp/kafkaconfig/start.sh', 'consumer', 'single']
-    log.info("Running %s", consumer_cmd)
-    consumer_output = sdk_tasks.task_exec(client, ' '.join(consumer_cmd))
-    log.info("Consumer output: %s", consumer_output)
-
-    assert str(message) in ' '.join(str(c) for c in consumer_output)
+    auth.wait_for_brokers(kafka_client["id"], kafka_client["brokers"])
+    auth.send_and_receive_message(kafka_client["id"])
